Added image repository, completed page layout configuration
I used a personal image as my placeholder for the hero section of the webpage and
added it into the new directory images. I then added the remaining flex styles in
order to get the webpage looking more similar to the template provided by TOP.

I struggled initially figuring out how to use flex positioning on an element that
was already a flex item, but using Chrome Devtools and some patience gave me a
better understanding of using flex.

I have also started using viewport height and width as values in order to create a
webpage that works when resized to different  viewport sizes, I haven't fixed every
element like this in the current webpage yet though.

I also looked up object-fit: cover; css style as a way of using css to crop an image
onto a webpage rather than change the aspect ratio when size is changed.